docs(DVLSForLinux): fix one-liner install command (#10)
We can’t pipe using `|` because our script is also reading user input
from stdin. When using `|`, we modify stdin to be the stdout of the
command coming before it. Instead, we fully read the file in memory,
and give the result to `bash` using the `<` operator. This is performed
in-line using parenthesis `(` `)` bash syntax to group statements
together. This doesn’t capture stdin, and allow the user to type as
usual.